+++ app/adapter/incoming/web/user.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends
 from starlette.responses import Response
 
 from app.adapter.incoming.web.schema.request.user_authorization import UserAuthorization
-from app.application.port.incoming.oauth_signin_use_case import OauthSignInUseCase
+from app.port.incoming.oauth_signin_use_case import OauthSignInUseCase
 from app.configs import Settings, get_settings
 from app.dependency.oauth_service import get_oauth_sign_in_service
