+++ app/adapter/outbound/user_create.py
@@ -2,10 +2,10 @@ from fastapi import Depends
 from sqlalchemy.exc import NoResultFound
 from sqlalchemy.orm import Session
 
-from app.application import entity
-from app.application.domain import model
-from app.application.port.incoming.oauth_signin_use_case import UserInfo
-from app.application.port.outbound.user_create import UserCreatePort
+from app import entity
+from app.domain import model
+from app.port.incoming.oauth_signin_use_case import UserInfo
+from app.port.outbound.user_create import UserCreatePort
 from app.dependency.db import get_db
