+++ app/entity/message.py
@@ -3,7 +3,7 @@ from datetime import datetime
 from sqlalchemy import Column, BigInteger, Integer, String, ForeignKey, Text, DateTime, func
 from sqlalchemy.orm import Mapped, relationship
 
-from app.application.entity import Base
+from app.entity import Base
 
 
 class Message(Base):
